shoping-agent/pyproject.toml
[project]
name = "shoping-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
]
